Add separate testable package
