+++ .gitignore
@@ -1,4 +1,4 @@
-/target
+target
 DawnCache
 GPUCache
 .DS_Store
