+++ tests/helper.rs
@@ -1,0 +1,4 @@
+// Run subprocess in a separate executable file
+fn main() {
+    wew::execute_subprocess();
+}
